Fix paths after the move which also fixes the tests
