+++ scripts/PH_SH_exporter.js
@@ -20,7 +20,7 @@
  * }}} */
 
 /* Required modules {{{ */
-var opening_hours = require('./opening_hours.js');
+var opening_hours = require('../build/opening_hours.js');
 var fs            = require('fs');
 var glob          = require('glob');
 var yaml          = require('js-yaml');
